fix: 修复区块渲染高度偏移问题

+++ src/common/network/sync/ChunkSync.hpp
@@ -52,7 +52,7 @@ public:
     static size_t calculateSectionSize(const ChunkSection& section);
 
     // 区块段位掩码计算
-    static u16 calculateSectionMask(const ChunkData& chunk);
+    static u32 calculateSectionMask(const ChunkData& chunk);
 };
 
 // ============================================================================
